+++ R/aaa.R
@@ -7,3 +7,6 @@ var_lab_mat <- ".m"
 
 dots_for_names <- ".."
 dots_for_names_expr <- "\\.\\."
+
+adjust_opts <- c(x_only="pad_x", from_y="from_y")
+